Move att-storage to NOT be codegenned (has no codegen input)
