+++ examples/all-clusters-minimal-app/linux/main-common.cpp
@@ -17,7 +17,7 @@
  */
 
 #include "include/tv-callbacks.h"
-#include <app-common/zap-generated/att-storage.h>
+#include <app/att-storage.h>
 #include <app-common/zap-generated/attributes/Accessors.h>
 #include <app/CommandHandler.h>
 #include <app/clusters/identify-server/identify-server.h>
