+++ examples/lighting-app/infineon/cyw30739/src/LightingManager.cpp
@@ -18,7 +18,7 @@
  */
 
 #include "LightingManager.h"
-#include <app-common/zap-generated/att-storage.h>
+#include <app/att-storage.h>
 #include <app-common/zap-generated/attributes/Accessors.h>
 #include <app/util/af.h>
 #include <stdio.h>
