+++ src/app/clusters/groups-server/groups-server.cpp
@@ -17,7 +17,7 @@
 
 #include "groups-server.h"
 
-#include <app-common/zap-generated/att-storage.h>
+#include <app/att-storage.h>
 #include <app-common/zap-generated/attributes/Accessors.h>
 #include <app-common/zap-generated/cluster-objects.h>
 #include <app-common/zap-generated/ids/Clusters.h>
