+++ src/app/clusters/mode-select-server/mode-select-server.cpp
@@ -15,7 +15,7 @@
  *    limitations under the License.
  */
 
-#include <app-common/zap-generated/att-storage.h>
+#include <app/att-storage.h>
 #include <app-common/zap-generated/attributes/Accessors.h>
 #include <app-common/zap-generated/cluster-objects.h>
 #include <app-common/zap-generated/ids/Attributes.h>
