+++ src/app/util/attribute-storage.h
@@ -26,7 +26,7 @@
 #include <platform/CHIPDeviceLayer.h>
 
 #if !defined(EMBER_SCRIPTED_TEST)
-#include <app-common/zap-generated/att-storage.h>
+#include <app/att-storage.h>
 #endif
 
 #if !defined(ATTRIBUTE_STORAGE_CONFIGURATION) && defined(EMBER_TEST)
