--- src/app/zap-templates/templates/app/att-storage.zapt
@@ -1,20 +1,0 @@
-{{> header}}
-
-// Prevent multiple inclusion
-#pragma once
-
-// Cluster masks modify how clusters are used by the framework
-//
-// Does this cluster have init function?
-#define CLUSTER_MASK_INIT_FUNCTION (0x01)
-// Does this cluster have attribute changed function?
-#define CLUSTER_MASK_ATTRIBUTE_CHANGED_FUNCTION (0x02)
-// Bits 0x04 and 0x08 are free.
-// Does this cluster have shutdown function?
-#define CLUSTER_MASK_SHUTDOWN_FUNCTION (0x10)
-// Does this cluster have pre-attribute changed function?
-#define CLUSTER_MASK_PRE_ATTRIBUTE_CHANGED_FUNCTION (0x20)
-// Cluster is a server
-#define CLUSTER_MASK_SERVER (0x40)
-// Cluster is a client
-#define CLUSTER_MASK_CLIENT (0x80)
