+++ src/darwin/Framework/CHIP/MTRIMDispatch.mm
@@ -18,7 +18,7 @@
 #import <Foundation/Foundation.h>
 
 #include <access/SubjectDescriptor.h>
-#include <app-common/zap-generated/att-storage.h>
+#include <app/att-storage.h>
 #include <app-common/zap-generated/callback.h>
 #include <app-common/zap-generated/cluster-objects.h>
 #include <app-common/zap-generated/ids/Attributes.h>
